Address web inspector review feedback
