+++ Sources/AnglesiteApp/FocusedSite.swift
@@ -10,17 +10,6 @@ extension FocusedValues {
     }
 }
 
-private struct FocusedPreviewKey: FocusedValueKey { typealias Value = PreviewModel }
-
-extension FocusedValues {
-    /// The focused site window's `PreviewModel`, published by `SiteWindow`. Lets View-menu commands
-    /// (e.g. `WebInspectorCommands`) reach the live preview without owning it.
-    var preview: PreviewModel? {
-        get { self[FocusedPreviewKey.self] }
-        set { self[FocusedPreviewKey.self] = newValue }
-    }
-}
-
 /// Must be `Commands` (not `App`) — `@FocusedValue` only tracks scene focus inside a `View`/`Commands` node.
 struct ExportSiteCommands: Commands {
     @FocusedValue(\.siteID) private var focusedSiteID
@@ -38,26 +27,6 @@ struct ExportSiteCommands: Commands {
                 }
             }
             .disabled(focusedSiteID == nil)
-        }
-    }
-}
-
-/// "Show Web Inspector" in the View menu — opens the focused site window's preview inspector
-/// (control-click already exposes WebKit's native "Inspect Element"). Enabled only when a site
-/// window is focused; a no-op until that window's dev server is ready and the web view exists.
-struct WebInspectorCommands: Commands {
-    @FocusedValue(\.preview) private var focusedPreview
-
-    var body: some Commands {
-        // `after: .toolbar` keeps "Show Web Inspector" adjacent to "Show Debug Pane" (which also uses
-        // `.toolbar`) — the two developer-tool toggles sit together (⌥⌘I next to ⌥⌘D). `.sidebar`
-        // would place it earlier, separated by the standard Toolbar entries.
-        CommandGroup(after: .toolbar) {
-            Button("Show Web Inspector") {
-                focusedPreview?.showWebInspector()
-            }
-            .keyboardShortcut("i", modifiers: [.command, .option])
-            .disabled(focusedPreview == nil)
         }
     }
 }
