+++ docs/build-plan.md
@@ -48,7 +48,7 @@ This is the riskiest piece, so do it first.
 
 ## Phase 4 — WKWebView + edit overlay (v0 core)
 
-1. ✅ `PreviewView` (`NSViewRepresentable<WKWebView>`, AnglesiteApp): reloads `http://localhost:<port>` whenever the URL changes (incl. when a supervised dev-server restart rebinds a new port — surfaced via `AstroDevServer.onReadyURLChange`). The testable core is `PreviewSession` (actor, AnglesiteCore): `start(siteID:siteDirectory:)` resolves how to run `astro dev` for a site (`node_modules/.bin/astro` via the vendored Node, or `.failed("dependencies not installed — run npm install")`), drives an `AstroDevServer`, and exposes `State ∈ {idle, starting, ready(url), failed(msg)}` plus an `observe()` change stream. `PreviewModel` (`@Observable`, AnglesiteApp) mirrors that state; `ContentView` is now a sidebar of discovered sites + a main pane that shows the live preview for the selected site (placeholder/error states otherwise) — per-site *windows* are still Phase 9. `WebViewBridge.localDevConfiguration()` (no disk cache in Debug) + `applyLocalDevDefaults` (`isInspectable` in Debug, so ⌥⌘I devtools work); `Info.plist` gained `NSAllowsLocalNetworking` for the plain-http localhost load. *Note:* end-to-end "preview a real site" needs the site's `node_modules` (i.e. `npm install` having run — see #6); until then the preview pane surfaces the "dependencies not installed" state and the tests use a `/bin/sh` fixture dev server.
+1. ✅ `PreviewView` (`NSViewRepresentable<WKWebView>`, AnglesiteApp): reloads `http://localhost:<port>` whenever the URL changes (incl. when a supervised dev-server restart rebinds a new port — surfaced via `AstroDevServer.onReadyURLChange`). The testable core is `PreviewSession` (actor, AnglesiteCore): `start(siteID:siteDirectory:)` resolves how to run `astro dev` for a site (`node_modules/.bin/astro` via the vendored Node, or `.failed("dependencies not installed — run npm install")`), drives an `AstroDevServer`, and exposes `State ∈ {idle, starting, ready(url), failed(msg)}` plus an `observe()` change stream. `PreviewModel` (`@Observable`, AnglesiteApp) mirrors that state; `ContentView` is now a sidebar of discovered sites + a main pane that shows the live preview for the selected site (placeholder/error states otherwise) — per-site *windows* are still Phase 9. `WebViewBridge.localDevConfiguration()` (no disk cache in Debug) + `applyPreviewDefaults` (`isInspectable` for preview inspection); `Info.plist` gained `NSAllowsLocalNetworking` for the plain-http localhost load. *Note:* end-to-end "preview a real site" needs the site's `node_modules` (i.e. `npm install` having run — see #6); until then the preview pane surfaces the "dependencies not installed" state and the tests use a `/bin/sh` fixture dev server.
 2. ✅ `WKScriptMessageHandler` for the `anglesite` namespace (#16). The JS → native channel lives in `AnglesiteBridge`: `EditMessage.decode(from:)` strictly validates `{id, type, path, selector, op, value}` at the boundary (only `type == "anglesite:apply-edit"` is accepted; unknown types fail decode); `EditRouter` is the seam (`LoggingEditRouter` is the wired default — logs to the Debug pane and replies `"Phase 5 lands the server side"`; `MCPApplyEditRouter` is the Phase-5-ready scaffold that calls `MCPClient.callTool(name: "anglesite:apply-edit", arguments: …)` and maps the result). `AnglesiteScriptHandler` (`WKScriptMessageHandler`) decodes, routes, and replies via `evaluateJavaScript("window.anglesite?._handleReply?.(<EditReply>)")` on the originating WKWebView; the decode → route → reply construction is factored into a pure `handle(body:via:)` so it's unit-tested without `WKScriptMessage`. `WebViewBridge.localDevConfiguration(handler:)` registers the handler on the user-content controller; `PreviewView` wires a per-webview `AnglesiteScriptHandler(router: LoggingEditRouter())`.
 3. ✅ JS edit overlay (#17). Lives at `JS/edit-overlay/` (TS source) and builds to a single `Resources/edit-overlay/overlay.js` IIFE via esbuild — `scripts/build-overlay.sh` runs as a prebuild phase (uses the vendored Node when present, system `npm` otherwise; idempotent on the `npm ci` step via `node_modules/.bin/esbuild` as the canary). Modules: `selector.ts` (pure `:nth-of-type` CSS path; first-cut strategy, #18 will finalize), `messages.ts` (post over `window.webkit.messageHandlers.anglesite`, await replies on `window.anglesite._handleReply`), `overlay.ts` (DOM behavior: hover outline / click → `contentEditable=true` / blur → post on change / drop image on `<img>` → post `set-image` with base64). Injected via `WebViewBridge.makeOverlayUserScript(in:)` → `WKUserScript` at `atDocumentEnd`, all frames; missing bundle is non-fatal (preview just loads without edit affordances). Tests via vitest + jsdom (19 in `JS/edit-overlay/test/` covering selector, messages, hover, click-to-edit, idempotent install) plus the Swift-side `WebViewBridgeTests` for the user-script loader. Drop-image behavior is implemented but not unit-tested (jsdom's DataTransfer/FileReader make it fiddly — manual smoke instead).
 4. ✅ **Selector strategy: server-side resolution (#18, decided 2026-05-13).** The overlay sends a structured `ElementInfo` payload — `{tag, id?, classes, nthChild, ancestors?, dataAnglesiteId?, dataTestId?, role?, ariaLabel?, textContent?}` matching the typedef in `anglesite/server/selector.mjs` — and the plugin invokes `buildSelector(info)` server-side. One source of truth for the priority order (`data-anglesite-id` > `data-testid` > `#id` > role/aria > stable classes > `tag:nth-child`), no fork-prone duplicated JS. `JS/edit-overlay/src/selector.ts` exposes `elementInfoFor(element)`; ancestors are root-first and stop at `<body>`. The bridge stays a relay: `EditMessage.selector` is `JSONValue` (validated as an object at decode), forwarded as-is by `MCPApplyEditRouter` to the `anglesite:apply-edit` tool when Phase 5 wires it up.
